Moved CUDA code to extension
