+++ Project.toml
@@ -4,8 +4,6 @@ authors = ["lmejn <33491793+lmejn@users.noreply.github.com>"]
 version = "0.13.6"
 
 [deps]
-Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 CoordinateTransformations = "150eb455-5306-5404-9cee-2592286d6298"
 DICOM = "a26e6606-dd52-5f6a-a97f-4f611373d757"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
@@ -28,6 +26,13 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
+
+[weakdeps]
+Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+
+[extensions]
+RoentgenCUDAExt = "CUDA"
 
 [compat]
 Adapt = "3"
