+++ src/Roentgen.jl
@@ -4,10 +4,6 @@ module Roentgen
 using SparseArrays
 using StaticArrays
 
-# CUDA
-using CUDA
-import Adapt
-
 # Coordinates
 using CoordinateTransformations
 using Rotations
